+++ packages/vorbis_rs/src/decoder/decoder_impl.rs
@@ -19,7 +19,7 @@ use crate::{common::VorbisError, decoder::VorbisAudioSamples};
 /// A decoder that turns a perceptually-encoded, non-chained Ogg Vorbis stream into
 /// blocks of planar, single-precision float audio samples.
 pub struct VorbisDecoder<R: Read> {
-	ogg_vorbis_file: OggVorbis_File,
+	ogg_vorbis_file: Box<OggVorbis_File>,
 	source: PhantomData<R>,
 	last_audio_block: Option<VorbisAudioSamples>
 }
@@ -33,13 +33,16 @@ impl<R: Read> VorbisDecoder<R> {
 	/// I/O errors that might happen during that operation will be returned to the
 	/// caller.
 	pub fn new<S: Into<Box<R>>>(source: S) -> Result<Self, VorbisError> {
-		let mut ogg_vorbis_file = MaybeUninit::uninit();
-
 		// The source read needs to be allocated in the heap (i.e., boxed) to have a
 		// constant memory address. Then leak it to a raw pointer to hand its ownership
 		// over to C code. Related, interesting read about trait objects and FFI:
 		// https://adventures.michaelfbryan.com/posts/ffi-safe-polymorphism-in-rust/
 		let source = Box::into_raw(source.into());
+
+		// The underlying `OggVorbis_File` struct also needs to have a constant memory address because
+		// it stores the vorbis_dsp_state and vorbis_block states by value, and the latter stores a
+		// pointer to the former that would be invalidated on move
+		let mut ogg_vorbis_file = Box::new_uninit();
 
 		// SAFETY: we assume ov_open_callbacks follows its documented contract
 		unsafe {
@@ -120,7 +123,7 @@ impl<R: Read> VorbisDecoder<R> {
 		// VorbisAudioSamples implementation for more safety information
 		unsafe {
 			let samples_read = vorbisfile_return_value_to_result!(ov_read_float(
-				&mut self.ogg_vorbis_file,
+				&mut *self.ogg_vorbis_file,
 				sample_buf.as_mut_ptr(),
 				2048, // Most stereo Ogg Vorbis files in the wild use a maximum block size of 2048 samples
 				current_bitstream.as_mut_ptr()
@@ -164,7 +167,7 @@ impl<R: Read> VorbisDecoder<R> {
 
 impl<R: Read> Drop for VorbisDecoder<R> {
 	fn drop(&mut self) {
-		unsafe { ov_clear(&mut self.ogg_vorbis_file) };
+		unsafe { ov_clear(&mut *self.ogg_vorbis_file) };
 	}
 }